backend/requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
google-generativeai==0.8.3
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
